Create dir before saving settings: case of locale and keyboard
If the path to a settings file does not exist, then
the file is not created, and an error is returned. This is
not hte same behavior as in systemd-localed. So we create
the directory before saving. Here in the case of the locale
and keyboard settings (save machinery is common to both).