Add a "finish" label in callbacks
When a call back finishes without error, or if there is an
error but we don't use g_dbus_method_invocation_xxxerror,
we need to call blocaled_locale1_complete_xxx. Since
the code may come from various places, add a "finish" label
before this call. We only use a goto once, but it may be handy
to have it if we add further code paths.