fix: set correct locale for testing under musl (and glibc)
Some tests are failing when run on a musl system. This is due to the
fact that some functions of glib library use unicode characters in their
output. As glibc and musl use different default locales if setlocale is
not correctly setup in the program, it leads to different character
outputs. glibs replaces unicode characters with ? where musl replaces
them with *. This was the cause of the failing tests.
This commit sets a correct default locale that make glibc and musl able
to display the correct unicode characters.