Bump minimal glib version
We now use errno without including <errno.h>. This is possible because
glib.h includes gtestutils.h, which includes errno.h. But this was added
in glib-2.61. So we should have this version as minimal. We could of
course include errno.h ourselves, but we do not have ways to test older
versions of glib anyway. And glib-2.61 is almost five years old, so
I don't think it is a big requirement.