Use files() function instead of absolute path in test dir
For tests, we recompile ùain.c with different config.h. The source is
in src, so we had "meson.project_source_root() / 'src/main.c'" as the
source. It is better to use the files function to obtain a "file"
object that can be used in any directory, letting meson do the work
of defining paths to it.